making hpx compile with less warnings
git-svn-id: https://svn.cct.lsu.edu/repos/projects/parallex/trunk/hpx@6550 5079c8e4-9419-0410-93bb-9e5ee278c886

+++ src/util/parse_command_line.cpp
@@ -373,7 +373,7 @@ namespace hpx { namespace util
             argv[i] = const_cast<char*>(args[i].c_str());
 
         return parse_commandline(
-            app_options, args.size(), argv.get(), vm, mode, visible);
+            app_options, static_cast<int>(args.size()), argv.get(), vm, mode, visible);
     }
 
     ///////////////////////////////////////////////////////////////////////////
